SMS-CAS/requirements.txt
Flask
mysql-connector-python
confluent-kafka
kafka-python
prometheus-client
